Penambahan fitur no image

+++ SITA Project/Delete_User.php
@@ -7,10 +7,23 @@ $databasePassword = '';
 $conn = mysqli_connect($databaseHost, $databaseUsername, $databasePassword, $databaseName);
 
 $nik = $_GET['nik'];
+
+$sql = "SELECT * FROM user WHERE nik=$nik";
+$result = $conn->query($sql);
+$row = $result->fetch_assoc();
+
 $result = mysqli_query($conn, "DELETE FROM user WHERE nik=$nik");
 
-$file = $_GET['img'];
-unlink($file);
+if("no_img.jpg"==$row['link_profile_pic'])
+{
+	header("Location:List_User.php");
+}
+else
+{
+	echo $row['link_profile_pic'];
+	$file = $_GET['img'];
+	unlink($file);
+	header("Location:List_User.php");
+}
 
-header("Location:List_User.php");
 ?>
